V1

+++ src/tasks/services/tasks.service.ts
@@ -1,33 +1,42 @@
 import {
-  ForbiddenException,
-  Injectable,
-  NotFoundException,
+  Inject,
+  Injectable, Logger,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Task } from '../domain/entities/task.entity';
-import { CreateTaskDto } from '../application/dtos/create-task.dto';
-import { UpdateTaskDto } from '../application/dtos/update-task.dto';
-import { lastValueFrom } from 'rxjs';
-import { ExternalTaskApiService } from '../infrastructure/clients/external-task-api.service';
-import { ExternalTaskDto } from '../application/dtos/external-task.dto';
+import { Task } from '../entities/task.entity';
+import { CreateTaskDto } from '../dtos/create-task.dto';
+import { UpdateTaskDto } from '../dtos/update-task.dto';
+import { ExternalTaskDto } from '../dtos/external-task.dto';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import { PaginationQueryDto } from '../../shared/dtos/pagination-query.dto';
 import { User, UserRole } from '../../users/entities/user.entity';
-import { ForbiddenTaskAccessException } from '../domain/errors/forbidden-access.exception';
-import { TaskNotFoundException } from '../domain/errors/task-not-found.exception';
+import { ForbiddenTaskAccessException } from '../errors/forbidden-access.exception';
+import { TaskNotFoundException } from '../errors/task-not-found.exception';
 import { UsersService } from '../../users/users.service';
+import { ITaskClient, ITaskClientName } from '../clients/task-client.interface';
+import { SyncTasksResponseDto } from '../dtos/sync-tasks-response.dto';
+import { ConfigService } from '@nestjs/config';
 
 export type RequestingUser = Pick<User, 'id' | 'role'>;
 
 @Injectable()
 export class TasksService {
+  protected readonly logger = new Logger(TasksService.name);
+
   constructor(
+    private configService: ConfigService,
     @InjectRepository(Task)
     private tasksRepo: Repository<Task>,
     private eventEmitter: EventEmitter2,
     private readonly userService: UsersService,
+    @Inject(ITaskClientName)
+    private readonly taskClient: ITaskClient,
   ) {}
+
+  async getExternalTasks(): Promise<ExternalTaskDto[]> {
+    return this.taskClient.findAll();
+  }
 
   async createTask(
     createDto: CreateTaskDto,
@@ -38,7 +47,8 @@ export class TasksService {
 
     const newTask = this.tasksRepo.create({
       title: createDto.title,
-      userId: createDto.userId,
+      assignedUserId: createDto.userId,
+      createdByUserId: requestingUser.id,
       completed: false,
       createdAt: new Date(),
       updatedAt: new Date(),
@@ -49,23 +59,23 @@ export class TasksService {
 
   async deleteTask(
     taskId: string,
-    userId: string,
+    userId: number,
     requestingUser: RequestingUser,
   ): Promise<boolean> {
     this.ensureTaskOwnershipOrAdmin(requestingUser, userId);
     await this.findUserTaskByIdOrFail(taskId, userId);
-    await this.tasksRepo.delete({ id: taskId, userId });
+    await this.tasksRepo.delete({ id: taskId, assignedUserId: userId });
     return true;
   }
 
   async getUserTasks(
-    userId: string,
+    userId: number,
     getDto: PaginationQueryDto,
   ): Promise<{ data: Task[]; total: number }> {
     const { page, pageSize } = getDto;
     const skip = (page - 1) * pageSize;
 
-    const whereConditions = { userId };
+    const whereConditions = { assignedUserId: userId };
 
     const [data, total] = await this.tasksRepo.findAndCount({
       where: whereConditions,
@@ -74,7 +84,7 @@ export class TasksService {
       order: {
         createdAt: 'DESC',
       },
-      cache: 60000 * 10,
+      cache: this.configService.get<number>('CACHE_TTL_MS') || 300000
     });
 
     return {
@@ -83,14 +93,14 @@ export class TasksService {
     };
   }
 
-  async findUserTaskById(taskId: string, userId: string): Promise<Task | null> {
+  async findUserTaskById(taskId: string, userId: number): Promise<Task | null> {
     return this.tasksRepo.findOne({
-      where: { id: taskId, userId: userId },
+      where: { id: taskId, assignedUserId: userId },
       cache: 60000 * 10,
     });
   }
 
-  async findUserTaskByIdOrFail(taskId: string, userId: string): Promise<Task> {
+  async findUserTaskByIdOrFail(taskId: string, userId: number): Promise<Task> {
     const task = await this.findUserTaskById(taskId, userId);
 
     if (!task) {
@@ -100,30 +110,38 @@ export class TasksService {
     return task;
   }
 
-  async syncExternalTasks(): Promise<void> {
-    /*
-    const { data: externalTasks } = await lastValueFrom(
-      this.externalTaskApi.findAll(),
-    );
+  async syncExternalTasks(): Promise<SyncTasksResponseDto> {
+    const externalTasks= await this.getExternalTasks();
 
+    //move to batches to increase performance
+    let syncedTasks = 0;
     for (const externalTask of externalTasks) {
-      const exists = await this.tasksRepo.findOne({
-        where: {
-          title: externalTask.title,
-          userId: String(externalTask.userId),
-        },
-      });
 
-      if (!exists) {
-        const mapped = this.mapExternalToLocalTask(externalTask);
-        const task = this.tasksRepo.create(mapped);
-        const saved = await this.tasksRepo.save(task);
+      const userExists = await this.userService.findOne({ id: externalTask.userId });
 
-        this.eventEmitter.emit('TASK_CREATED', saved);
+      if(userExists){
+        const taskExists = await this.tasksRepo.findOne({
+          where: {
+            title: externalTask.title,
+            assignedUserId: externalTask.userId,
+          },
+        });
+
+        if (!taskExists) {
+          const mapped = this.mapExternalToLocalTask(externalTask);
+          const task = this.tasksRepo.create(mapped);
+          const saved = await this.tasksRepo.save(task);
+
+          this.eventEmitter.emit('TASK_CREATED', saved);
+          syncedTasks++;
+        }
       }
     }
 
-     */
+    return {
+      processedTasks: externalTasks.length,
+      syncedTasks: syncedTasks,
+    };
   }
 
   async updateTask(
@@ -155,7 +173,7 @@ export class TasksService {
 
   private ensureTaskOwnershipOrAdmin(
     requesting: RequestingUser,
-    targetUserId: string,
+    targetUserId: number,
   ): void {
     if (requesting.role === UserRole.USER && requesting.id !== targetUserId) {
       throw new ForbiddenTaskAccessException();
@@ -164,7 +182,8 @@ export class TasksService {
 
   private mapExternalToLocalTask(dto: ExternalTaskDto): Partial<Task> {
     return {
-      userId: String(dto.userId),
+      assignedUserId: dto.userId,
+      createdByUserId: dto.userId,
       title: dto.title,
       completed: dto.completed,
       createdAt: new Date(),
